Added Group reactions map

+++ BloodBankData/dbo/Tables/BloodGroup.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[BloodGroup]
 (
-	[Id] INT NOT NULL PRIMARY KEY, 
+	[Id] TINYINT NOT NULL PRIMARY KEY, 
     [GroupRh] CHAR(10) NOT NULL, 
 	/*Maps Group to the it's respective grading.*/
